feat: fetchos 커스텀 에러 클래스 구현 및 fetchInstance 에러 처리 (#311)
